refactor IncidentMap
