--- src/signals/IncidentMap/components/FilterCategoryPanel/index.ts
@@ -1,2 +1,0 @@
-export { FilterCategoryPanel } from './FilterCategoryPanel'
-export type { FilterCategory } from './FilterCategoryPanel'
