+++ src/signals/IncidentMap/components/IncidentMap/styled.ts
@@ -3,7 +3,7 @@
 
 import styled from 'styled-components'
 import Map from 'components/Map'
-import { breakpoint, Button, themeSpacing } from '@amsterdam/asc-ui'
+import { breakpoint } from '@amsterdam/asc-ui'
 
 export const Wrapper = styled.div`
   position: absolute;
@@ -23,6 +23,7 @@ export const Wrapper = styled.div`
 export const Container = styled.div`
   position: relative;
   display: flex;
+  flex-direction: row-reverse;
   width: 100%;
   height: 100%;
   @media screen and ${breakpoint('max-width', 'tabletM')} {
@@ -34,30 +35,4 @@ export const StyledMap = styled(Map)`
   height: 100%;
   width: 100%;
   z-index: 0;
-`
-
-export const StyledButton = styled(Button)`
-  position: absolute;
-  top: ${themeSpacing(5)};
-  left: calc(33% - ${themeSpacing(2)});
-  z-index: 3;
-  width: ${themeSpacing(9)};
-  box-shadow: ${themeSpacing(1)} ${themeSpacing(1)} ${themeSpacing(1)}
-    rgba(0, 0, 0, 0.1);
-  &.hiddenPanel {
-    left: 0;
-  }
-  @media screen and ${breakpoint('max-width', 'tabletM')} {
-    transform: rotate(-90deg);
-    top: calc(50% - ${themeSpacing(3)});
-    left: calc(50% - 18px);
-    box-shadow: ${themeSpacing(0)} ${themeSpacing(0)} ${themeSpacing(0)}
-      rgba(0, 0, 0, 0.1);
-    &.hiddenPanel {
-      left: calc(50% - 18px);
-      top: calc(100% - ${themeSpacing(11)});
-      box-shadow: ${themeSpacing(1)} ${themeSpacing(1)} ${themeSpacing(1)}
-        rgba(0, 0, 0, 0.1);
-    }
-  }
 `
