+++ src/signals/IncidentMap/components/utils/index.ts
@@ -1,0 +1,1 @@
+export { updateFilterCategories } from './update-filter-categories'
